Add PublishFileStorage and WantFileStorage
