+++ src/Omnius.Xeus.Service/Primitives/IStorage.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Omnius.Core.Cryptography;
+
+namespace Omnius.Xeus.Service.Primitives
+{
+    public interface IStorage : IReadOnlyStorage
+    {
+        ValueTask WriteAsync(OmniHash rootHash, OmniHash targetHash, ReadOnlyMemory<byte> memory, CancellationToken cancellationToken = default);
+    }
+}
